+++ core/ui/src/main/kotlin/com/afternote/core/ui/Label.kt
@@ -1,4 +1,4 @@
-package com.afternote.core.ui.form
+package com.afternote.core.ui
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Box
