+++ core/ui/src/main/kotlin/com/afternote/core/ui/ProfileImage.kt
@@ -1,4 +1,4 @@
-package com.afternote.core.ui.form
+package com.afternote.core.ui
 
 import android.util.Log
 import androidx.annotation.DrawableRes
@@ -25,7 +25,6 @@ import coil3.compose.AsyncImagePainter
 import coil3.network.NetworkHeaders
 import coil3.network.httpHeaders
 import coil3.request.ImageRequest
-import com.afternote.core.ui.R
 import com.afternote.core.ui.theme.AfternoteTheme
 
 private const val TAG = "ProfileImage"
