+++ core/ui/src/main/kotlin/com/afternote/core/ui/RecordQnAListItem.kt
@@ -1,4 +1,4 @@
-package com.afternote.core.ui.list
+package com.afternote.core.ui
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Arrangement
