+++ core/ui/src/main/kotlin/com/afternote/core/ui/popup/AfternotePopupCardLayout.kt
@@ -1,4 +1,4 @@
-package com.afternote.core.ui.feedback
+package com.afternote.core.ui.popup
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Arrangement
@@ -20,7 +20,7 @@ import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import com.afternote.core.ui.expand.dropShadow
+import com.afternote.core.ui.dropShadow
 import com.afternote.core.ui.theme.AfternoteDesign
 import com.afternote.core.ui.theme.AfternoteTheme
