+++ core/ui/src/main/kotlin/com/afternote/core/ui/popup/ConfirmationPopup.kt
@@ -1,4 +1,4 @@
-package com.afternote.core.ui.feedback
+package com.afternote.core.ui.popup
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.clickable
@@ -22,7 +22,7 @@ import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.window.Dialog
 import androidx.compose.ui.window.DialogProperties
-import com.afternote.core.ui.expand.dropShadow
+import com.afternote.core.ui.dropShadow
 import com.afternote.core.ui.theme.AfternoteDesign
 import com.afternote.core.ui.theme.AfternoteTheme
