+++ core/ui/src/main/kotlin/com/afternote/core/ui/popup/InfoPopup.kt
@@ -1,4 +1,4 @@
-package com.afternote.core.ui.feedback
+package com.afternote.core.ui.popup
 
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.fillMaxWidth
@@ -15,7 +15,7 @@ import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.window.Dialog
-import com.afternote.core.ui.expand.dropShadow
+import com.afternote.core.ui.dropShadow
 import com.afternote.core.ui.theme.AfternoteDesign
 import com.afternote.core.ui.theme.AfternoteTheme
