+++ feature/afternote/presentation/src/main/kotlin/com/afternote/feature/afternote/presentation/author/home/AfternoteHomeScreen.kt
@@ -7,7 +7,7 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.tooling.preview.Preview
-import com.afternote.core.ui.scaffold.PlusButtonFAB
+import com.afternote.core.ui.scaffold.FAB.PlusButtonFAB
 import com.afternote.core.ui.scaffold.topbar.DetailTopBar
 import com.afternote.core.ui.scaffold.topbar.HomeTopBar
 import com.afternote.core.ui.theme.AfternoteTheme
