+++ feature/onboarding/presentation/src/main/java/com/afternote/feature/onboarding/presentation/signup/SignUpScreen.kt
@@ -32,9 +32,9 @@ import androidx.compose.ui.text.input.ImeAction
 import androidx.compose.ui.text.input.KeyboardType
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import com.afternote.core.ui.form.AfternoteTextField
-import com.afternote.core.ui.form.PasswordMaskTransformation
 import com.afternote.core.ui.scaffold.topbar.DetailTopBar
+import com.afternote.core.ui.textfield.AfternoteTextField
+import com.afternote.core.ui.textfield.PasswordMaskTransformation
 import com.afternote.core.ui.theme.AfternoteDesign
 import com.afternote.core.ui.theme.AfternoteTheme
 import com.afternote.feature.onboarding.presentation.R
